+++ code/src/hooks/geocode/useExistingLocationCoordinates.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
-import { AgentResponseBody } from "types/backend-agent";
-import { makeInternalRegistryAPIwithParams } from "utils/internal-api-services";
+import { AgentResponseBody, InternalApiIdentifierMap } from "types/backend-agent";
+import { makeInternalRegistryAPIwithParams, queryInternalApi } from "utils/internal-api-services";
 
 export interface ExistingCoordinatesDescriptor {
   isFetching: boolean;
@@ -23,14 +23,9 @@ export function useExistingLocationCoordinates(
   useEffect(() => {
     // Declare an async function to get geocoordinates associated with the location
     const getGeoCoordinates = async (location: string): Promise<void> => {
-      const res = await fetch(
-        makeInternalRegistryAPIwithParams("geodecode", location),
-        {
-          cache: "no-store",
-          credentials: "same-origin",
-        }
+      const resBody: AgentResponseBody = await queryInternalApi(
+        makeInternalRegistryAPIwithParams(InternalApiIdentifierMap.GEODECODE, location)
       );
-      const resBody: AgentResponseBody = await res.json();
       const existingCoordinates: number[] = (
         resBody.data?.items as Record<string, unknown>[]
       )?.[0]?.coordinates as number[];